feat: refactored build-metadata
