--- contracts/release1/build1/build-metadata.json
@@ -1,9 +1,0 @@
-{
-  "ui": "",
-  "change": "Initial build.",
-  "pluginSetupABI": {
-    "prepareInstallation": ["uint256 number"],
-    "prepareUpdate": [],
-    "prepareUninstallation": []
-  }
-}
