--- contracts/release1/build2/build-metadata.json
@@ -1,11 +1,0 @@
-{
-  "ui": "",
-  "change": "Added a `address account` as a second state variable.",
-  "pluginSetupABI": {
-    "prepareInstallation": ["uint256 number", "address account"],
-    "prepareUpdate": {
-      "fromBuild1": ["address account"]
-    },
-    "prepareUninstallation": []
-  }
-}
