--- contracts/release1/build3/build-metadata.json
@@ -1,12 +1,0 @@
-{
-  "ui": "",
-  "change": "Replaced the `STORE_PERMISSION` permission by the more specific `STORE_NUMBER_PERMISSION` and `STORE_ACCOUNT_PERMISSION` and added the `NumberStored` and `AccountStored` events.",
-  "pluginSetupABI": {
-    "prepareInstallation": ["uint256 number", "address account"],
-    "prepareUpdate": {
-      "fromBuild1": ["address account"],
-      "fromBuild2": []
-    },
-    "prepareUninstallation": []
-  }
-}
